balanced stack

+++ stack.py
@@ -9,7 +9,8 @@ class Stack():
 
       if empty == 0 :
         self.stack = np.random.choice(range(100), size, replace=False)
+        self.stack = self.stack.astype('float32')
         self.top = self.stack[-1]
       else :
-        self.stack = np.full(size, np.nan)
+        self.stack = np.full(size, np.nan, dtype=np.float32)
         self.top = None
